add countdown
