+++ src/components/countdown/style.module.scss
@@ -1,0 +1,7 @@
+@use "../../styles/vars.scss" as v; // allows you to use pre-defined colors
+
+.date {
+  color: black;
+  font-size: 1.25rem;
+  font-weight: 700;
+}
